requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
distro==1.9.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.48.0
pydantic==2.9.2
pydantic_core==2.23.4
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.97
pywin32==306
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.30.6
